chatbot/pyproject.toml
[project]
name = "football-chatbot"
version = "0.1.0"
description = "AI Assistant for Football Online using LangGraph"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
